Added a prison level with patrolling wolves
The first wolf, on the right, is intended to walk forward and backwards along that corridor. You get past by waiting till it's walking backwards and run past it. The second wolf stands still. You can see it through the bars. You have to go around the pillar, get it's attention and then Cube so that it runs past you. Once it's back to searching behaviour, you uncube and run for the exit.